requirements.txt
requests~=2.26.0
pillow~=8.4.0